% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_access.R
\name{get_data}
\alias{get_data}
\alias{dataset_versions}
\alias{dataset_version_current}
\alias{dataset_del}
\title{Get leaf13C dataset}
\usage{
get_data(version = NULL, path = NULL)

dataset_versions(local = FALSE, path = NULL)

dataset_version_current(local = FALSE, path = NULL)

dataset_del(version, path = NULL)
}
\arguments{
\item{version}{Optional character string specifying the dataset version to retrieve.
If \code{NULL} (default), the most recent available version will be used.}

\item{path}{Optional file path to store and access the dataset locally.
If \code{NULL} (default), a system-appropriate location will be used,
determined via \code{rappdirs::user_data_dir("leaf13C")}.}

\item{local}{Logical indicating if local or github versions should
be polled.  With any luck, \code{local=FALSE} is a superset of
\code{local=TRUE}.  For \code{mydata_version_current}, if
\code{TRUE}, but there are no local versions, then we do check
for the most recent github version.}
}
\value{
A data frame containing the leaf13C dataset.
}
\description{
Retrieve a versioned copy of the leaf13C dataset
}
\details{
This function provides access to a versioned dataset hosted on GitHub
using the \pkg{datastorr} framework. It will download and cache the dataset
if not already available locally, and return it as a data frame.

If no version is specified, the most recent available version will be used.
If no path is specified, a suitable persistent directory will be chosen
automatically using \code{rappdirs::user_data_dir()}.
}
\examples{
# Load the most recent version of the dataset
data <- get_data()

# Load a specific version of the dataset
data <- get_data(version = "0.1.0")

# Use a custom path
data <- get_data(path = "~/my_data")
}
